requirements.txt
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
asgiref==3.8.1
attr==0.3.1
attrs==25.3.0
azure-core==1.33.0
azure-storage-blob==12.25.1
bleach==5.0.1
boto==2.49.0
boto3==1.37.38
botocore==1.37.38
boxing==0.1.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.3.0
expiringdict==1.2.2
filelock==3.16.1
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
label-studio==1.13.1
label-studio-sdk==1.0.5
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
openai==1.78.1
opencv-python==4.11.0.86
opentelemetry-api==1.33.0
ordered-set==4.0.2
packaging==25.0
pandas==2.0.3
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyRFC3339==2.0.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
python-pptx==1.0.2
pytz==2022.7.1
PyYAML==6.0.2
redis==3.5.3
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rq==1.10.1
rsa==4.9.1
rules==2.2
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.4
sentence-transformers==3.2.1
sentry-sdk==2.28.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.17.2
XlsxWriter==3.2.3
xmljson==0.2.1
zipp==3.20.2
